edit architecture and configuration
