+++ .gitignore
@@ -13,3 +13,8 @@
 .externalNativeBuild
 .cxx
 local.properties
+/gradle/
+/gradlew
+/gradlew.bat
+/settings.gradle
+/gradle.properties
