+++ app/src/main/java/com/example/dota2/MainActivity.kt
@@ -1,17 +1,10 @@
 package com.example.dota2
 
+import MainScreen
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Surface
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.SideEffect
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.core.view.WindowCompat
-import com.example.dota2.ui.theme.AppTheme
-import com.google.accompanist.systemuicontroller.rememberSystemUiController
 
 
 class MainActivity : ComponentActivity() {
@@ -23,32 +16,5 @@ class MainActivity : ComponentActivity() {
         setContent {
             MainScreen()
         }
-    }
-}
-
-
-@Composable
-fun MainScreen() {
-    AppTheme {
-
-        ApplySystemBarColors()
-
-        Surface(
-            color = AppTheme.BgColors.primary,
-            modifier = Modifier
-                .fillMaxSize(),
-        ) {
-            DotaScreen()
-        }
-    }
-}
-
-
-@Composable
-private fun ApplySystemBarColors() {
-    val systemUiController = rememberSystemUiController()
-    SideEffect {
-        systemUiController.setStatusBarColor(color = Color.Transparent)
-        systemUiController.setNavigationBarColor(color = Color.Transparent)
     }
 }
