+++ app/src/main/java/com/example/dota2/components/Rating.kt
@@ -1,4 +1,4 @@
-package com.example.dota2
+package com.example.dota2.components
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
@@ -6,10 +6,14 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
+import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.example.dota2.R
+import com.example.dota2.ui.theme.AppTheme
 
 
 @Composable
@@ -42,5 +46,22 @@ fun Rating(
                     .height(12.dp)
             )
         }
+    }
+}
+
+
+@Preview
+@Composable
+fun DotaRatingPreview() {
+    Surface(
+        color = AppTheme.BgColors.primary
+    ) {
+        Rating(
+            modifier = Modifier
+                .padding(start = 12.dp)
+                .height(14.dp)
+                .width(90.dp),
+            rating = 3.1f
+        )
     }
 }
