+++ app/src/main/java/com/example/dota2/components/RatingReview.kt
@@ -1,13 +1,17 @@
-package com.example.dota2
+package com.example.dota2.components
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.example.dota2.R
 import com.example.dota2.ui.theme.AppTheme
 
 
@@ -21,7 +25,7 @@ fun RatingReview(
         Text(
             text = rating.toString(),
             style = AppTheme.TextStyle.Bold_48,
-            color = AppTheme.TextColors.primary,
+            color = AppTheme.TextColors.secondary,
         )
         Column(
             modifier = Modifier.padding(
@@ -41,5 +45,25 @@ fun RatingReview(
                 color = AppTheme.TextColors.message,
             )
         }
+    }
+}
+
+
+@Preview
+@Composable
+fun DotaRatingReviewPreview() {
+    Surface(
+        color = AppTheme.BgColors.primary
+    ) {
+        RatingReview(
+            rating = 4.9f,
+            downloads = stringResource(id = R.string.downloads),
+            modifier = Modifier
+                .padding(
+                    start = 24.dp,
+                    top = 12.dp,
+                    bottom = 32.dp
+                )
+        )
     }
 }
