+++ app/src/main/java/com/example/dota2/ui/MainScreen.kt
@@ -1,0 +1,35 @@
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.SideEffect
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import com.example.dota2.ui.dotascreen.DotaScreen
+import com.example.dota2.ui.theme.AppTheme
+import com.google.accompanist.systemuicontroller.rememberSystemUiController
+
+@Composable
+fun MainScreen() {
+    AppTheme {
+
+        ApplySystemBarColors()
+
+        Surface(
+            color = AppTheme.BgColors.primary,
+            modifier = Modifier
+                .fillMaxSize(),
+        ) {
+            DotaScreen()
+        }
+    }
+}
+
+
+@Composable
+private fun ApplySystemBarColors() {
+    val systemUiController = rememberSystemUiController()
+    SideEffect {
+        systemUiController.setStatusBarColor(color = Color.Transparent)
+        systemUiController.setNavigationBarColor(color = Color.Transparent)
+    }
+}
