+++ libs.versions.toml
@@ -5,11 +5,8 @@ compose-bom = "2023.10.01"
 activity-compose = "1.8.0"
 runtime = "1.5.4"
 systemuicontroller = "0.30.1"
-agp = "8.1.1"
 kotlin = "1.8.10"
-junit = "4.13.2"
-junit-ext = "1.1.5"
-espresso = "3.5.1"
+composeCompiller = "1.4.3"
 
 [libraries]
 coreKtx = { group = "androidx.core", name = "core-ktx", version.ref = "core-ktx" }
@@ -20,6 +17,7 @@ composeRuntime = { group = "androidx.compose.runtime", name = "runtime", version
 composeUi = { group = "androidx.compose.ui", name = "ui"}
 composePreview = { group = "androidx.compose.ui", name = "ui-tooling-preview"}
 composeDebugPreview = { group = "androidx.compose.ui", name = "ui-tooling"}
+composeGraphics = { group = "androidx.compose.ui", name = "ui-graphics"}
 composeMaterial = { group = "androidx.compose.material", name = "material"}
 composeMaterial3 = { group = "androidx.compose.material3", name = "material3"}
 accomponistSystemUiController = {group = "com.google.accompanist", name = "accompanist-systemuicontroller", version.ref = "systemuicontroller"}
